pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spark-databricks-utility-hub"
version = "0.1.0"
description = "PySpark + Databricks project with reusable JAR utilities"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Data Engineering Team" }
]
dependencies = [
  "pyspark==3.5.1",
  "delta-spark==3.2.0",
  "PyYAML==6.0.2",
  "python-json-logger==2.0.7",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==5.0.0",
  "ruff==0.11.2",
  "mypy==1.15.0",
  "types-PyYAML",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.mypy]
python_version = "3.11"
pretty = true
ignore_missing_imports = true
warn_unused_configs = true
